wip: fully migrate to wrapped, native ticket model
